refactor(backend): extract workflow definition assembler
